% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley_k.R
\name{ripley_k}
\alias{ripley_k}
\title{Ripley K function for 1D data}
\usage{
ripley_k(x, t = NA, xmin = NA, xmax = NA, weighting = "reflect", nstep = NA)
}
\arguments{
\item{x}{vector of values for analysis}

\item{t}{vector of windows}

\item{xmin}{Minimum possible for data x. If xmin is set to NA then the value is
calculated from x}

\item{xmax}{Maximum possible value for data x. If xmax i set to NA then the value is
calculcated from x}

\item{weighting}{A character string representing the type of weighting to apply
to mitigate against edge effects. Allowed values are "none" for no weighting, "reflect" to
increase the weight of points near xmin and xmax. "cycle" should added in thr future.}

\item{nstep}{An integer value used to automate the creation of vector of t values}
}
\description{
Uses Ripley K function to analyse whether data are evenly dispersed, random or clustered
}
