% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley_k_mc.R
\name{ripley_k_mc}
\alias{ripley_k_mc}
\title{Monte Carlo calculation for Ripley K function for 1D data}
\usage{
ripley_k_mc(n, t, xmin, xmax, edge, nsim)
}
\arguments{
\item{xmin}{Minimum possible for data x.}

\item{xmax}{Maximum possible value for data x.}

\item{edge}{A character string representing the type of edge effect compensation
to apply. The allowed values are "none" for no compensation, "reflect" to
increase the weight of points near xmin and xmax, and "periodic" to treat the
values as if they repeat periodically}

\item{nsim}{Number of simulations to run}
}
\description{
Uses Ripley K function to analyse whether data are evenly dispersed,
random or clustered @param n number of values for analysis @param t vector of windows
}
