% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley_k_score.R
\name{ripley_k_score}
\alias{ripley_k_score}
\title{Ripley K score}
\usage{
ripley_k_score(df, method = "sum")
}
\arguments{
\item{df}{Data frame generated by ripley_k}

\item{method}{Method to use to get the statistic. So far, only "sum"
(sum of differences between k and T) and sumabs (sum of absolute differences between k and T)
are valid}
}
\description{
Get a single value to summary the results from ripley_k
}
