% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley_k_mc.R
\name{ripley_k_mc}
\alias{ripley_k_mc}
\title{Monte Carlo calculation for Ripley K function for 1D data}
\usage{
ripley_k_mc(n, t, xmin, xmax, weighting = "reflect", nsim = 100)
}
\arguments{
\item{xmin}{Minimum possible for data x.}

\item{xmax}{Maximum possible value for data x.}

\item{weighting}{A character string representing the type of weighting to apply
to mitigate against edge effects. Allowed values are "none" for no weighting, "reflect" to
increase the weight of points near xmin and xmax. "cycle" should added in thr future.}

\item{nsim}{Number of simulations to run}
}
\description{
Uses Ripley K function to analyse whether data are evenly dispersed,
random or clustered @param n number of values for analysis @param t vector of windows
}
