% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley_k_stat.R
\name{ripley_k_stat}
\alias{ripley_k_stat}
\title{Statistics for Ripley K function in 1D}
\usage{
ripley_k_stat(
  x,
  t = NA,
  xmin = NA,
  xmax = NA,
  edge = "weight",
  nstep = NA,
  nsim = 100
)
}
\arguments{
\item{x}{vector of values for analysis}

\item{t}{vector of windows}

\item{xmin}{Minimum possible for data x. If xmin is set to NA then the value is
calculated from x}

\item{xmax}{Maximum possible value for data x. If xmax i set to NA then the value is
calculcated from x}

\item{edge}{A character string representing the type of edge effect compensation
to apply. The allowed values are "none" for no compensation, "weight" to
increase the weight of points near xmin and xmax, and "periodic" to treat the
values as if they repeat periodically}

\item{nstep}{An integer value used to automate the creation of vector of t values}
}
\description{
This takes a dataset, calculates K and makes a comparison
with a comparable set of Monte Carlo simulations to
estimate significance
}
